requirements.txt
torch==2.1.0
torchvision==0.16.0
streamdiffusion
ndi-python
numpy<2
transformers<4.36
Pillow
opencv-python
